refactor: layouts styled and structure
