+++ components/layouts/auth/index.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import styled from "styled-components";
 
 export const AuthCont = styled.main`
@@ -6,3 +7,7 @@ export const AuthCont = styled.main`
   display: flex;
   align-items: center;
 `;
+
+export default function AuthLayout({ children }) {
+  return <AuthCont>{children}</AuthCont>;
+}
