--- layouts/auth/index.jsx
@@ -1,6 +1,0 @@
-import React from "react";
-import { AuthCont } from "./styled";
-
-export default function AuthLayout({ children }) {
-  return <AuthCont>{children}</AuthCont>;
-}
