+++ pages/_app.js
@@ -1,6 +1,6 @@
 import { Provider } from "react-redux";
 import store from "../app/store";
-import MainLayout from "../layouts/main";
+import MainLayout from "../components/layouts/main";
 import GlobalStyles from "../theme/global";
 import "../styles/index.css";
