+++ pages/api/hello.js
@@ -1,0 +1,3 @@
+export default function handler(req, res) {
+  return res.status(200).json({ hello: "World" });
+}
